py_sphyxer-master/requirements.txt

matplotlib
numpy
pandas
pmdarima
PyYAML
sphinx
statsmodel
yahoofinancials
